requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
waitress==2.1.2
WebOb==1.8.7
whitenoise==6.6.0
